cards/requirements.txt
django==4.1.7
django-simple-history==3.2.0
djangorestframework==3.14.0
drf_yasg==1.21.5
schedule==1.1.0
django-filter==3.5.2

